SideNav Implemented

+++ src/components/Customer/DisableCustomer.js
@@ -3,23 +3,28 @@ import { useHistory } from "react-router-dom";
 const DisableCustomer = () => {
   const history = useHistory();
   const DisableCustomerAccount = (e) => {
-    deleteCustomerService()
-      .then((response) => {
-        console.log(response.data);
-        localStorage.removeItem("loggedInUser");
-        history.push("/");
-        window.location.reload(true);
-      })
-      .catch(() => {
-        alert("Error Occured");
-      });
+    var proceed = window.confirm("Your Account will be permenantly deleted. Are you sure you want to proceed?");
+    if (proceed) {
+      deleteCustomerService()
+        .then((response) => {
+          console.log(response.data);
+          localStorage.removeItem("loggedInUser");
+          history.push("/");
+          window.location.reload(true);
+        })
+        .catch(() => {
+          alert("Error Occured");
+        });
+    } else {
+    }
+
   };
 
   return (
     <input
       type="submit"
-      className="btn btn-success mt-3"
-      value="Disable Account"
+      className="btn btn-danger mt-3 w-100"
+      value="Disable/Delete Account"
       onClick={DisableCustomerAccount}
     />
   );
